+++ rand_core/CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### API changes
 - Relax `Sized` bound on impls of `SeedableRng` (#1641)
-- Fix `OsError::raw_os_error` on UEFI targets by returning `Option<usize>` (#1665)
 - Move `rand_core::impls::*` to `rand_core::le` module (#1667)
 - Use Edition 2024 and MSRV 1.85 (#1668)
 - Remove fn `TryRngCore::read_adapter(..) -> RngReadAdapter` (replaced with `rand::RngReader`) (#1669)
+- Remove feature `os_rng`, structs `OsRng` and `OsError` and fns `from_os_rng`, `try_from_os_rng` (#1674)
+- Remove feature `std` (#1674)
+- Removed dependency `getrandom` (#1674)
 
 ## [0.9.3] — 2025-02-29
 ### Other